add evaluation rule logic.

+++ crawler/src/main/java/data/CandleStick.java
@@ -14,9 +14,9 @@ public class CandleStick {
     private String symbol;
     private long startTime;
     private long finishTime;
-    private long close;
-    private long open;
-    private long low;
-    private long high;
+    private double close;
+    private double open;
+    private double low;
+    private double high;
 
 }
